LayoutTests/imported/w3c/web-platform-tests/pointerevents/pointerevent_releasepointercapture_onpointercancel_touch-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Pointer Events Capture Test - release capture on pointercancel

Test Description: This test checks if setCapture/releaseCapture functions works properly. Complete the following actions:
Touch black rectangle and do not release your touch
Move your touch to scroll the page. "lostpointercapture" should be logged inside of the black rectangle immediately after "pointercancel"

Test passes if the proper behavior of the events is observed.
Pointer Events Capture Test

Test complete: Scroll to Summary to view Pass/Fail Results.

The following pointer types were detected: (none).

The following events were logged: (none).


Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN pointer capture is released on pointercancel
